requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM API Interfaces
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Database
redis>=4.5.0

# Web Scraping & Search
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs>=3.0.0

# Data Processing & Validation
pydantic>=2.0.0

# Utilities
typing-extensions>=4.0.0
uuid>=1.30

# Development & Testing (Optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0